requirements.txt
PyYaml
bs4
requests